Update CUI
CUIの機能が一通り完成した！

+++ firmware/main_sensor_test/Inc/GNSS.h
@@ -36,7 +36,7 @@ struct gnss_goal {
 
 void init_gnss();
 void set_gnssGoal(const struct gnss_goal *data);
-void get_gnssGoal(int32_t *latitude, int32_t *longitude, uint64_t *dist);
+void get_gnssGoal(int32_t *latitude, int32_t *longitude, uint32_t *dist);
 int get_gnss(struct gnss* data, const uint32_t TIMEOUT);
 
 #endif /* GNSS_H_ */
